+++ src/p5/config_windowjs_p5.cc
@@ -1,3 +1,3 @@
-#include "config.h"
+#include "../config.h"
 
 bool windowjs_config_pass_args_to_loader = true;
